backend/requirements.txt
fastapi
uvicorn[standard]
Pillow
pillow-heif
python-multipart
openai
PyJWT
passlib
